% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depcens_audit.R
\name{is.depcens_audit}
\alias{is.depcens_audit}
\title{Test for DepCens-Audit objects}
\usage{
is.depcens_audit(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if \code{x} inherits from class \code{depcens_audit}.
}
\description{
Test for DepCens-Audit objects
}
